requirements.txt
depthai==2.5.0.0
Flask==2.0.1
Flask-JWT-Extended==4.2.1
Flask-Login==0.5.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
jsonschema==3.2.0
numpy==1.17.4
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.54
pandas==1.2.4
python-dateutil==2.7.3
SQLAlchemy==1.4.17
PyMySQL==1.0.2
paho-mqtt==1.5.1